log de administracion de usuarios

+++ src/UNO/EvaluacionesBundle/Controller/API/APIAdminController.php
@@ -8,7 +8,8 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\JsonResponse;
-
+use UNO\EvaluacionesBundle\Entity\LogDeleteReset;
+use UNO\EvaluacionesBundle\Entity\Userhttpsession;
 use UNO\EvaluacionesBundle\Controller\API\APIUtils;
 
 /**
@@ -31,6 +32,11 @@ class APIAdminController extends Controller{
     public function deletePersonAction(Request $request){
 
         $personId= $request->request->get('personId');
+        $user= $request->request->get('user');
+        $name= $request->request->get('name');
+        $schoolId= $request->request->get('schoolId');
+        $school= $request->request->get('school');
+        $executioner= $request->request->get('executioner');
 
         $em = $this->getDoctrine()->getManager();
         $PersonId = $em->getRepository('UNOEvaluacionesBundle:Person')->findOneBy(array('personid'=> $personId));
@@ -38,6 +44,7 @@ class APIAdminController extends Controller{
         if($PersonId) {
             $em->remove($PersonId);
             $em->flush();
+            $this->log(array('action' => 'Delete', 'personId' => $personId, 'user' => $user, 'name' => $name, 'schoolId' => $schoolId, 'school' => $school, 'surveyId' => null, 'survey' => null, 'executioner' => $executioner));
             return new JsonResponse(array('status' => true),200);
         }else{
             return new JsonResponse(array('status' => false),400);
@@ -51,16 +58,33 @@ class APIAdminController extends Controller{
     public function deleteSchoolAction(Request $request){
 
         $schoolId= $request->request->get('schoolId');
-       
-        $em = $this->getDoctrine()->getManager();
-        $school = $em->getRepository('UNOEvaluacionesBundle:School')->findOneBy(array('schoolid'=> $schoolId));
+        $schoolName= $request->request->get('school');
+        $executioner= $request->request->get('executioner');
 
-        if($school) {
-            $em->remove($school);
-            $em->flush();
-            return new JsonResponse(array('status' => true),200);
-        }else{
-            return new JsonResponse(array('status' => false),400);
+        $result = $this->executeQuery(
+            "DELETE P
+            FROM School S
+            inner join PersonSchool PS on S.schoolId = PS.schoolId
+            inner join Person P on PS.PersonId = P.personId
+            where S.schoolId = $schoolId
+            "
+        );
+
+        if($result) {
+            $em = $this->getDoctrine()->getManager();
+            $school = $em->getRepository('UNOEvaluacionesBundle:School')->findOneBy(array('schoolid' => $schoolId));
+
+            if ($school) {
+                $em->remove($school);
+                $em->flush();
+
+                $this->log(array('action' => 'Delete', 'personId' => null, 'user' => null, 'name' => null, 'schoolId' => $schoolId, 'school' => $schoolName, 'surveyId' => null, 'survey' => null, 'executioner' => $executioner));
+                return new JsonResponse(array('status' => true), 200);
+            } else {
+                return new JsonResponse(array('status' => false), 400);
+            }
+        }else {
+            return new JsonResponse(array('status' => false), 400);
         }
     }
 
@@ -71,6 +95,11 @@ class APIAdminController extends Controller{
     public function resetPersonAction(Request $request){
 
         $personId= $request->request->get('personId');
+        $user= $request->request->get('user');
+        $name= $request->request->get('name');
+        $schoolId= $request->request->get('schoolId');
+        $school= $request->request->get('school');
+        $executioner= $request->request->get('executioner');
 
         $result = $this->executeQuery(
             "DELETE A
@@ -90,6 +119,7 @@ class APIAdminController extends Controller{
             );
 
             if($result2) {
+                $this->log(array('action' => 'Reset Stats', 'personId' => $personId, 'user' => $user, 'name' => $name, 'schoolId' => $schoolId, 'school' => $school, 'surveyId' => null, 'survey' => null,'executioner' => $executioner));
                 return new JsonResponse(array('status' => true),200);
             }else{
                 return new JsonResponse(array('status' => false),400);
@@ -105,8 +135,13 @@ class APIAdminController extends Controller{
     public function resetPersonSurveyAction(Request $request){
 
         $personId= $request->request->get('personId');
+        $user= $request->request->get('user');
+        $name= $request->request->get('name');
+        $schoolId= $request->request->get('schoolId');
+        $school= $request->request->get('school');
         $surveyId= $request->request->get('surveyId');
-
+        $survey= $request->request->get('survey');
+        $executioner= $request->request->get('executioner');
 
         $result = $this->executeQuery(
             "DELETE A
@@ -128,6 +163,7 @@ class APIAdminController extends Controller{
             );
 
             if($result2) {
+                $this->log(array('action' => 'Reset Stats', 'personId' => $personId, 'user' => $user, 'name' => $name, 'schoolId' => $schoolId, 'school' => $school, 'surveyId' => $surveyId, 'survey' => $survey, 'executioner' => $executioner));
                 return new JsonResponse(array('status' => true),200);
             }else{
                 return new JsonResponse(array('status' => false),400);
@@ -142,7 +178,9 @@ class APIAdminController extends Controller{
      */
     public function resetSchoolAction(Request $request){
 
-        $schoolId= $request->request->get('schoolId');
+        $schoolId = $request->request->get('schoolId');
+        $school = $request->request->get('school');
+        $executioner = $request->request->get('executioner');
 
         $result = $this->executeQuery(
             "DELETE A
@@ -176,6 +214,7 @@ class APIAdminController extends Controller{
             );
 
             if($result2) {
+                $this->log(array('action' => 'Reset Stats', 'personId' => null, 'user' => null, 'name' => null, 'schoolId' => $schoolId, 'school' => $school, 'surveyId' => null, 'survey' => null, 'executioner' => $executioner));
                 return new JsonResponse(array('status' => true),200);
             }else{
                 return new JsonResponse(array('status' => false),400);
@@ -189,9 +228,11 @@ class APIAdminController extends Controller{
      */
     public function resetSchoolSurveyAction(Request $request){
 
-        $schoolId= $request->request->get('schoolId');
-        $surveyId= $request->request->get('surveyId');
-
+        $schoolId = $request->request->get('schoolId');
+        $school = $request->request->get('school');
+        $surveyId = $request->request->get('surveyId');
+        $survey = $request->request->get('survey');
+        $executioner = $request->request->get('executioner');
 
         $result = $this->executeQuery(
             "DELETE A
@@ -233,6 +274,7 @@ class APIAdminController extends Controller{
             );
 
             if($result2) {
+                $this->log(array('action' => 'Reset Stats', 'personId' => null, 'user' => null, 'name' => null, 'schoolId' => $schoolId, 'school' => $school, 'surveyId' => $surveyId, 'survey' => $survey, 'executioner' => $executioner));
                 return new JsonResponse(array('status' => true),200);
             }else{
                 return new JsonResponse(array('status' => false),400);
@@ -250,13 +292,33 @@ class APIAdminController extends Controller{
             $statement = $em->getConnection()->prepare($query);
             $statement->execute();
             $em->getConnection()->commit();
-
             return true;
 
         }catch (\PDOException $e) {
                 throw new \RuntimeException(sprintf('PDOException was thrown when trying to delete expired sessions: %s', $e->getMessage()), 0, $e);
         }
+    }
 
+    private function log($log){
 
+        $LogDeleteReset = new LogDeleteReset();
+
+        $LogDeleteReset->setAction($log["action"]);
+        $LogDeleteReset->setPersonId($log["personId"]);
+        $LogDeleteReset->setUser($log["user"]);
+        $LogDeleteReset->setName($log["name"]);
+        $LogDeleteReset->setSchoolId($log["schoolId"]);
+        $LogDeleteReset->setSchool($log["school"]);
+        $LogDeleteReset->setSurveyId($log["surveyId"]);
+        $LogDeleteReset->setSurvey($log["survey"]);
+        $LogDeleteReset->setDateAction(new \DateTime());
+        $LogDeleteReset->setIpCliente($_SERVER['REMOTE_ADDR']);
+        $LogDeleteReset->setExecutioner($log["executioner"]);
+
+        $em = $this->getDoctrine()->getManager();
+        // tells Doctrine you want to (eventually) save the Product (no queries yet)
+        $em->persist($LogDeleteReset);
+        // actually executes the queries (i.e. the INSERT query)
+        $em->flush();
     }
 }
